Add a label to display placeholder text.
This change adds a UILabel above the internalTextView,
configured to display placeholder text.  The label is
positioned such that the cursor of the textView is still
visible when the textView becomes firstResponder.

When the textView changes, the 'hidden' state of the
placeholder label is updated to show/hide the placeholder.